agents/research_assistant/requirements.txt
annotated-types==0.7.0
anthropic==0.68.1
anyio==4.11.0
async-timeout==4.0.3
beautifulsoup4==4.14.0
certifi==2025.8.3
charset-normalizer==3.4.3
distro==1.9.0
docstring_parser==0.17.0
exceptiongroup==1.3.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.11.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-anthropic==0.3.20
langchain-core==0.3.76
langchain-openai==0.3.33
langchain-text-splitters==0.3.11
langsmith==0.4.31
openai==1.109.1
orjson==3.11.3
packaging==25.0
pydantic==2.11.9
pydantic_core==2.33.2
python-dotenv==1.1.1
PyYAML==6.0.3
regex==2025.9.18
requests==2.32.5
requests-toolbelt==1.0.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.43
tenacity==9.1.2
tiktoken==0.11.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
wikipedia==1.4.0
zstandard==0.25.0
